[#1] 코드 정리
