--- AppName/appNameMessage/AddChallengeView.swift
@@ -1,21 +1,0 @@
-//
-//  AddChallenge.swift
-//  AppName
-//
-//  Created by hanseoyoung on 10/15/24.
-//
-
-import UIKit
-
-class AddChallengeView: UIViewController {
-
-    @IBAction func endTap(_ sender: Any) {
-        // 추후 동작 수정 필요 -> 일단은 네비게이션 이동으로 연결해둠
-//        guard let uvc = self.storyboard?.instantiateViewController(withIdentifier: "ChallengeRecord") else {
-//            return
-//        }
-//
-//        self.navigationController?.pushViewController(uvc, animated: true)
-    }
-
-}
