+++ AppName/appNameMessage/Model/ElecChallengeModel.swift
