+++ AppName/appNameMessage/View/AddChallengeView.swift
@@ -1,0 +1,16 @@
+//
+//  AddChallenge.swift
+//  AppName
+//
+//  Created by hanseoyoung on 10/15/24.
+//
+
+import UIKit
+
+class AddChallengeView: UIViewController {
+
+    @IBAction func endTap(_ sender: Any) {
+
+    }
+
+}
