+++ AppName/appNameMessage/View/ChallengeRecordTableViewCell.swift
@@ -9,21 +9,13 @@ import UIKit
 
 class ChallengeRecordTableViewCell: UITableViewCell {
 
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
 
-        // Configure the view for the selected state
     }
 
     @IBOutlet weak var recordTitle: UILabel!
-
     @IBOutlet weak var recordKwh: UILabel!
-    
     @IBOutlet weak var recordTarget: UILabel!
 
     func configure(with record: ChallengeRecord) {
