+++ AppName/appNameMessage/View/MessagesViewController.swift
@@ -8,30 +8,8 @@
 import UIKit
 import Messages
 
-class MessagesViewController: MSMessagesAppViewController, UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return sampleData.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "ChallengeCollectionCell", for: indexPath) as! ChallengeCollectionCell
+class MessagesViewController: MSMessagesAppViewController {
 
-                let challenge = sampleData[indexPath.item]
-                cell.configure(challenge)
-
-                return cell
-    }
-
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let padding: CGFloat = 8 // 셀 사이의 간격
-        let collectionViewSize = collectionView.frame.size.width - padding * 4 
-
-        // 셀의 너비를 3개로 나누어서 계산
-        let cellWidth = collectionViewSize / 3
-        return CGSize(width: cellWidth, height: cellWidth) // 셀을 정사각형으로 설정
-    }
-
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
@@ -94,17 +72,32 @@ class MessagesViewController: MSMessagesAppViewController, UICollectionViewDeleg
 
 }
 
-class ChallengeCollectionCell: UICollectionViewCell {
 
-    @IBOutlet weak var challengeTitle: UILabel!
-    @IBOutlet weak var challengeKwh: UILabel!
-    @IBOutlet weak var challengeDday: UILabel!
+// MARK: - CollectionView
 
-    func configure(_ elec: ElecChallenge) {
-        challengeTitle.text = "\(elec.name)"
-        challengeKwh.text = "\(elec.totalKwh)kWh"
-        challengeDday.text = "D-\(elec.date)"
+extension MessagesViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return sampleData.count
     }
-}
 
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "ChallengeCollectionCell", for: indexPath) as! ChallengeCollectionCell
+
+                let challenge = sampleData[indexPath.item]
+                cell.configure(challenge)
+
+                return cell
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let padding: CGFloat = 8 // 셀 사이의 간격
+        let collectionViewSize = collectionView.frame.size.width - padding * 4
+
+        // 셀의 너비를 3개로 나누어서 계산
+        let cellWidth = collectionViewSize / 3
+        return CGSize(width: cellWidth, height: cellWidth) // 셀을 정사각형으로 설정
+    }
+
+}
